commit msg is required when deploying the blog

+++ deploy_blog.sh
@@ -1,13 +1,17 @@
 #!/bin/sh
+if [ $# -eq 0 ]; then
+    echo "git need a commit message!"
+    exit 1
+fi
 cd ~/raw.gongshw.github.io; 
 jekyll build || error_exit "$LINENO: jekyll build failed";
 git add .;
-git commit -am "Latest build: ${1}";
+git commit -am "${1}";
 git push;
 rm -rf ../gongshw.github.io/*;
 cp -r _site/* ../gongshw.github.io/;
 cd ../gongshw.github.io/;
 git add .;
-git commit -am "Latest build: ${1}";
+git commit -am "${1}";
 git push;
